kernel module porting -isEmpty to size() == 0 -map Float to float? in .NET -cast Integer, Float, Boolean to int, float, boolean -other minor improvements

+++ itextsharp.io/itextsharp/io/font/otf/ActualTextIterator.cs
@@ -109,7 +109,7 @@ namespace iTextSharp.IO.Font.Otf
 					break;
 				}
 				// TODO zero glyph is a special case. Unicode might be special
-				toUnicodeMapResult.Append(TextUtil.ConvertFromUtf32(currentGlyph.GetUnicode()));
+				toUnicodeMapResult.Append(TextUtil.ConvertFromUtf32((int) currentGlyph.GetUnicode()));
 			}
 			return needsActualText || !toUnicodeMapResult.ToString().Equals(glyphLinePart.actualText
 				);
